@@ -8,17 +8,22 @@
         "ext-ctype": "*",
         "ext-iconv": "*",
         "composer/package-versions-deprecated": "1.11.99.4",
-        "doctrine/annotations": "^1.13",
+        "doctrine/annotations": "1.14.x-dev",
         "doctrine/doctrine-bundle": "^2.4",
         "doctrine/doctrine-migrations-bundle": "^3.1",
         "doctrine/orm": "^2.9",
+        "easycorp/easyadmin-bundle": "3.0.x-dev",
         "symfony/console": "5.3.*",
         "symfony/dotenv": "5.3.*",
         "symfony/flex": "^1.3.1",
         "symfony/framework-bundle": "5.3.*",
         "symfony/proxy-manager-bridge": "5.3.*",
         "symfony/runtime": "5.3.*",
-        "symfony/yaml": "5.3.*"
+        "symfony/yaml": "5.3.*",
+        "symfony/security-core": "5.4.x-dev",
+        "symfony/security-http": "5.4.x-dev",
+        "symfony/security-bundle": "5.4.x-dev",
+        "symfony/validator": "5.4.x-dev"
     },
     "require-dev": {
         "symfony/debug-bundle": "5.3.*",
