@@ -6,7 +6,9 @@
     use Doctrine\ORM\Mapping as ORM;
     use Symfony\Component\Validator\Constraints as Assert;
 
-	trait AttachmentsTrait
+
+
+    trait AttachmentsTrait
 	{
         /**
          * @var string
