@@ -3,7 +3,7 @@
 	namespace App\Entity\Module;
 
 	use Doctrine\ORM\Mapping as ORM;
-    use Symfony\Component\Validator\Constraints as Assert;
+    
 
     /**
 	 * Modules
